Move KernelBoot functions and struct to own file
